# src/main/resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pricing API",
    "description": "API for managing product prices by brand on a specific date.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api"
    }
  ],
  "paths": {
    "/prices": {
      "get": {
        "summary": "Get the price of a product for a brand on a specific date",
        "operationId": "getProductPriceOfBrandOnDated",
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "required": true,
            "description": "ID of the product",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "brandId",
            "in": "query",
            "required": true,
            "description": "ID of the brand",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "priceDate",
            "in": "query",
            "required": true,
            "description": "Date and time for applying the price (ISO-8601)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Price found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceResponse"
                }
              }
            }
          },
          "404": {
            "description": "No price found for the given request"
          }
        },
        "tags": [
          "Prices"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "PriceRequest": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the product"
          },
          "brandId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the brand"
          },
          "priceDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time for applying the price (ISO-8601)"
          }
        },
        "required": ["productId", "brandId", "priceDate"]
      },
      "PriceResponse": {
        "type": "object",
        "properties": {
          "priceList": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the price list"
          },
          "productId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the product"
          },
          "brandId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the brand"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Start date and time of the price (ISO-8601)"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "End date and time of the price (ISO-8601)"
          },
          "value": {
            "type": "number",
            "format": "float",
            "description": "Value of the price"
          }
        }
      }
    }
  }
}